Arbitrum Nitro v3.2.1

WIP: currently running into same error as the CI when running `make
build-replay-env`:

    GOOS=wasip1 GOARCH=wasm go build -o target/machines/latest/replay.wasm ./cmd/replay/...
    target/bin/prover target/machines/latest/replay.wasm --generate-binaries target/machines/latest \
    -l target/machines/latest/forward.wasm -l target/machines/latest/soft-float.wasm -l target/machines/latest/wasi_stub.wasm -l target/machines/latest/host_io.wasm -l target/machines/latest/user_host.wasm -l target/machines/latest/arbcompress.wasm -l target/machines/latest/program_exec.wasm -l target/machines/latest/espresso_crypto.wasm
    Error: failed to validate WASM binary at "target/machines/latest/wasi_stub.wasm"

    Caused by:
    0: failed to validate target/machines/latest/wasi_stub.wasm
    1: reference-types not enabled: zero byte expected (at offset 0x18d8)

    Location:
        prover/src/binary.rs:329:10
    make: *** [Makefile:436: target/machines/latest/machine.wavm.br] Error 1

+++ arbnode/dataposter/slice/slicestorage.go
@@ -89,8 +89,8 @@ func (s *Storage) Put(_ context.Context, index uint64, prev, new *storage.Queued
 		}
 		s.queue = append(s.queue, newEnc)
 	} else if index >= s.firstNonce {
-		queueIdx := int(index - s.firstNonce)
-		if queueIdx > len(s.queue) {
+		queueIdx := index - s.firstNonce
+		if queueIdx > uint64(len(s.queue)) {
 			return fmt.Errorf("attempted to set out-of-bounds index %v in queue starting at %v of length %v", index, s.firstNonce, len(s.queue))
 		}
 		prevEnc, err := s.encDec().Encode(prev)
